+++ src/tvheadend/AutoRecordings.cpp
@@ -55,7 +55,7 @@ void AutoRecordings::SyncDvrCompleted()
 
 int AutoRecordings::GetAutorecTimerCount() const
 {
-  return m_autoRecordings.size();
+  return static_cast<int>(m_autoRecordings.size());
 }
 
 void AutoRecordings::GetAutorecTimers(std::vector<kodi::addon::PVRTimer>& timers)
