+++ src/tvheadend/TimeRecordings.cpp
@@ -47,7 +47,7 @@ void TimeRecordings::SyncDvrCompleted()
 
 int TimeRecordings::GetTimerecTimerCount() const
 {
-  return m_timeRecordings.size();
+  return static_cast<int>(m_timeRecordings.size());
 }
 
 void TimeRecordings::GetTimerecTimers(std::vector<kodi::addon::PVRTimer>& timers)
